requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
boto3==1.10.20
botocore==1.13.20
certifi==2019.9.11
chardet==3.0.4
Deprecated==1.2.7
docutils==0.15.2
idna==2.8
importlib-metadata==0.23
jmespath==0.9.4
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
py==1.8.0
PyGithub==1.44.1
PyJWT==1.7.1
pyparsing==2.4.5
pytest==5.2.4
python-dateutil==2.8.0
requests==2.22.0
s3transfer==0.2.1
six==1.13.0
urllib3==1.25.7
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.6.0
